requirements.txt
youtube-transcript-api
google-generativeai
tqdm
deepl
regex
